DDPB-2359: Renames SubmissionCsvFilter to ReportSubmissionDownloadFilterDto
